games-strategy/spring: Fix -104 ebuild
Emerging with the 'java' use flag enabled broke, because
java-depend-query was not able to infer the correct Java JDK version
from the ":*" slotted virtual/jdk declaration. We know depend on JDK
1.8 (Java 8) as this it the stabilized JDK version in the Gentoo
ebuild repository and sufficient for spring.

Also sort keywords, update header and remove unnecessary set of 'S' in
case we are the live the ebuild.

This ebuild now passes "repoman full".

Package-Manager: Portage-2.3.51, Repoman-2.3.11
Signed-off-by: Florian Schmaus <flo@geekplace.eu>